+++ src/services/sync/telemetry/sampling.ts
@@ -38,14 +38,10 @@ export const syncSentryBeforeSendTransaction: ReturnsUndefined<NonNullable<Sentr
 
 // sentry doesn't bother to expose your chosen environment in tracesSampler
 // so we have to make consumers pass in our greedy option
-export const createSyncSentryTracesSampler = (greedy = false) => {
+export const createSyncSentryTracesSampler = () => {
   const sampler: ReturnsUndefined<NonNullable<SentryBrowserOptions['tracesSampler']>> = (context) => {
     if (!context.name.startsWith(SYNC_TELEMETRY_TRACE_PREFIX)) {
       return undefined
-    }
-
-    if (greedy) {
-      return true
     }
 
     const { parentSampled, attributes } = context
@@ -58,7 +54,7 @@ export const createSyncSentryTracesSampler = (greedy = false) => {
       return userBucketedSampler(attributes.userId as string | number, 0.1)
     }
 
-    return false
+    return undefined
   }
 
   return sampler